move log viewer component
