+++ web-app/src/app/client/modules/log-viewer/log-text-field/log-text-field.component.css
