+++ web-app/src/app/client/modules/log-viewer/log-text-field/log-text-field.component.ts
@@ -3,7 +3,7 @@ import { debounceTime } from 'rxjs/operators';
 import { fromEvent, Observable } from 'rxjs';
 import { select, Store } from '@ngrx/store';
 
-import * as fromRoot from '../../../app.reducers';
+import * as fromRoot from '../../../../app.reducers';
 import * as fromLogFile from '../store/log-file.reducers';
 
 @Component({
