+++ web-app/src/app/client/modules/log-viewer/log-tool-bar/log-tool-bar.component.html
