+++ web-app/src/app/client/modules/log-viewer/log-tool-bar/log-tool-bar.component.ts
@@ -1,20 +1,17 @@
 import { Component, OnInit } from '@angular/core';
 import * as logAction from '../store/log-file.actions';
 import { Store } from '@ngrx/store';
-import * as fromRoot from '../../../app.reducers';
+import * as fromRoot from '../../../../app.reducers';
 
 @Component({
   selector: 'app-log-tool-bar',
   templateUrl: './log-tool-bar.component.html',
-  styleUrls: ['./log-tool-bar.component.css']
+  styleUrls: ['./log-tool-bar.component.css'],
 })
 export class LogToolBarComponent implements OnInit {
+  constructor(private store: Store<fromRoot.State>) {}
 
-  constructor(private store: Store<fromRoot.State>) {
-  }
-
-  ngOnInit() {
-  }
+  ngOnInit() {}
 
   clearLog() {
     this.store.dispatch(logAction.clearLog());
@@ -23,5 +20,4 @@ export class LogToolBarComponent implements OnInit {
   sendTestMessage() {
     this.store.dispatch(logAction.sendTestMessage());
   }
-
 }
