+++ web-app/src/app/client/modules/log-viewer/log-viewer/log-viewer.component.css
