+++ web-app/src/app/client/modules/log-viewer/store/log-file.effects.ts
